+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 - Enh #350: Adapt to `Like` changes in `yiisoft/db` package (@vjik)
 - Enh #352: Support column's collation (@Tigrov)
 - New #358: Add `Connection::getColumnBuilderClass()` method (@Tigrov)
+- New #357: Implement `ArrayMergeBuilder`, `LongestBuilder` and `ShortestBuilder` classes (@Tigrov)
 
 ## 1.3.0 March 21, 2024